Added the new triangulation algorithm, fixed floor depth calculation. created a volume calculation test system.
Former-commit-id: 856735d3df332e41ce93e0d9d4c40927c1d602da
Former-commit-id: 15b8a033b375fa33f0d74f9d61cc0a2b61462125

+++ DepthMapProcessor/Structures.h
@@ -7,9 +7,9 @@ typedef unsigned int uint;
 
 struct ObjDimDescription
 {
-	short Width;
-	short Height;
-	short Depth;
+	int Width;
+	int Height;
+	int Depth;
 };
 
 struct RelPoint
